+++ apps-rs/test/src/lib.rs
@@ -22,11 +22,13 @@
 //!  - Input and output serialization.
 //!  - Default smart contract functionalities.
 //!  - Trigger exceptional conditions.
-
+//!
+use random::Source;
 use serde_derive::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::time::SystemTime;
 use std::{
     alloc::{alloc, Layout},
-    collections::HashMap,
     convert::TryInto,
     mem::align_of,
 };
@@ -34,7 +36,6 @@ use trinci_sdk::{
     rmp_deserialize, rmp_serialize, tai::AssetTransferArgs, value, AppContext, PackedValue, Value,
     WasmError, WasmResult,
 };
-
 trinci_sdk::app_export!(
     // Input and output serialization.
     echo_generic,
@@ -45,13 +46,22 @@ trinci_sdk::app_export!(
     balance,
     transfer,
     notify,
+    // Utility methods
+    mint,
+    store_data,
+    // Host function tests
+    get_account_keys,
     // Trigger exceptional conditions.
     divide_by_zero,
     trigger_panic,
     exhaust_memory,
     infinite_recursion,
     infinite_loop,
-    null_pointer_indirection
+    null_pointer_indirection,
+    // Deterministic contract
+    get_random_sequence,
+    get_hashmap,
+    get_time
 );
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -71,6 +81,14 @@ struct EchoArgs<'a> {
     pub vec8: Vec<u8>,
     pub vec16: Vec<u16>,
     pub map: HashMap<&'a str, SubStruct<'a>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+#[cfg_attr(test, derive(PartialEq, Clone, Default))]
+struct StoreDataArgs<'a> {
+    pub key: &'a str,
+    #[serde(with = "serde_bytes")]
+    pub data: Vec<u8>,
 }
 
 /// Returns the input data "as is".
@@ -116,6 +134,26 @@ fn balance(ctx: AppContext, _args: PackedValue) -> WasmResult<u64> {
     trinci_sdk::log("Called method `balance`");
     let value = load_my_asset(ctx.caller);
     Ok(value)
+}
+
+/// Mint some `account asset` units on the account
+fn mint(ctx: AppContext, args: u64) -> WasmResult<u64> {
+    trinci_sdk::log("Called method `mint`");
+
+    let value = load_my_asset(ctx.owner);
+    store_my_asset(ctx.owner, value + args);
+
+    Ok(value + args)
+}
+
+/// Store the data with the given key in the current account
+fn store_data(_ctx: AppContext, args: StoreDataArgs) -> WasmResult<()> {
+    trinci_sdk::store_account_data_mp!(args.key, &args.data)
+}
+
+/// Call the host function hf_get_data_keys
+fn get_account_keys(_ctx: AppContext, pattern: &str) -> WasmResult<Vec<String>> {
+    trinci_sdk::get_data_keys(pattern)
 }
 
 /// Call the host function hf_transfer
@@ -190,7 +228,7 @@ fn exhaust_memory(_ctx: AppContext, _args: Value) -> WasmResult<Value> {
 fn infinite_recursion(ctx: AppContext, args: Value) -> WasmResult<Value> {
     if let Value::Bool(first_call) = args {
         if first_call {
-            trinci_sdk::log("Called method `null_pointer_indirection`");
+            trinci_sdk::log("Called method `infinite_recursion`");
         }
         return infinite_recursion(ctx, value!(false));
     }
@@ -226,8 +264,44 @@ fn notify(ctx: AppContext, data: Value) -> WasmResult<()> {
     Ok(())
 }
 
+/// Return a random sequence (shall be deterministic)
+fn get_random_sequence(_ctx: AppContext, _args: PackedValue) -> WasmResult<PackedValue> {
+    trinci_sdk::log("Called method `random_sequence`");
+
+    let mut source = random::default();
+    let vector = source.iter().take(3).collect::<Vec<u64>>();
+
+    let buf = trinci_sdk::rmp_serialize(&vector)?;
+    Ok(PackedValue(buf))
+}
+
+/// Return an hashmap (shall be deterministic)
+fn get_hashmap(_ctx: AppContext, _args: PackedValue) -> WasmResult<PackedValue> {
+    trinci_sdk::log("Called method `return_hashmap`");
+
+    let mut hashmap: HashMap<&str, u64> = HashMap::default();
+
+    hashmap.insert("val1", 123);
+    hashmap.insert("val2", 456);
+    hashmap.insert("val3", 789);
+
+    let buf = trinci_sdk::rmp_serialize(&hashmap)?;
+    Ok(PackedValue(buf))
+}
+
+/// Try to access to system time.
+fn get_time(_ctx: AppContext, _args: PackedValue) -> WasmResult<u64> {
+    trinci_sdk::log("Called method `get_time`");
+
+    let sys_time = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap();
+    Ok(sys_time.as_secs())
+}
+
 #[cfg(test)]
 mod tests {
+
     use super::*;
     use trinci_sdk::not_wasm;
 
@@ -260,7 +334,7 @@ mod tests {
     }
 
     fn create_typed_echo_args() -> EchoArgs<'static> {
-        let mut map = HashMap::new();
+        let mut map = HashMap::default();
         map.insert(
             "k1",
             SubStruct {
@@ -392,5 +466,138 @@ mod tests {
         let res = not_wasm::call_wrap(notify, ctx, args).unwrap();
 
         assert_eq!(res, ());
+    }
+
+    #[test]
+    fn store_data_test() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = StoreDataArgs {
+            key: "foo",
+            data: vec![1, 2, 3],
+        };
+        let res = not_wasm::call_wrap(store_data, ctx, args).unwrap();
+
+        assert_eq!(res, ());
+        let buf = not_wasm::get_account_data(OWNER_ID, "foo");
+
+        let data: Vec<u8> = rmp_deserialize(&buf).unwrap();
+        let expected = vec![1, 2, 3];
+
+        assert_eq!(data, expected);
+    }
+
+    #[test]
+    fn get_keys_with_empty_pattern() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = "";
+        let err = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap_err();
+
+        assert_eq!(err.to_string(), "last char of search pattern must be '*'");
+    }
+
+    #[test]
+    fn get_keys_with_invalid_pattern_1() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = "x";
+        let err = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap_err();
+
+        assert_eq!(err.to_string(), "last char of search pattern must be '*'");
+    }
+
+    #[test]
+    fn get_keys_with_invalid_pattern_2() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = "abc:*sdfsdf";
+        let err = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap_err();
+
+        assert_eq!(err.to_string(), "last char of search pattern must be '*'");
+    }
+
+    #[test]
+    fn get_all_keys_not_existent() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = "*";
+        let res = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap();
+
+        assert_eq!(res, Vec::<String>::new());
+    }
+
+    #[test]
+    fn get_keys_not_existent() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        let args = "key:*";
+        let res = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap();
+
+        assert_eq!(res, Vec::<String>::new());
+    }
+
+    #[test]
+    fn get_keys_with_pattern() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        not_wasm::set_account_data(OWNER_ID, "abc", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc1", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc*:123", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "xyz", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "ab", &[1, 2, 3]);
+
+        let args = "abc*";
+        let mut res = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap();
+        res.sort();
+        let mut expected = vec![
+            "abc".to_string(),
+            "abc1".to_string(),
+            "abc*:123".to_string(),
+        ];
+        expected.sort();
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn get_keys_with_wildcard() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        not_wasm::set_account_data(OWNER_ID, "abc", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc1", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc*:123", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "xyz", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "ab", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "*", &[1, 2, 3]);
+
+        let args = "*";
+        let mut res = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap();
+        res.sort();
+        let mut expected = vec![
+            "abc".to_string(),
+            "abc1".to_string(),
+            "abc*:123".to_string(),
+            "xyz".to_string(),
+            "ab".to_string(),
+            "*".to_string(),
+        ];
+        expected.sort();
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn get_keys_with_wrong_pattern() {
+        let ctx = not_wasm::create_app_context(OWNER_ID, CALLER_ID);
+
+        not_wasm::set_account_data(OWNER_ID, "abc", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc1", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "abc*:123", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "xyz", &[1, 2, 3]);
+        not_wasm::set_account_data(OWNER_ID, "ab", &[1, 2, 3]);
+
+        let args = "hello*";
+        let res = not_wasm::call_wrap(get_account_keys, ctx, args).unwrap();
+
+        assert_eq!(res, Vec::<String>::new());
     }
 }